PUBDEV-676: More WIP towards row weights for DL.

+++ h2o-algos/src/main/java/hex/gram/Gram.java
@@ -850,7 +850,7 @@ public final class Gram extends Iced<Gram> {
       _gram.addRow(r, w);
       ++_nobs;
     }
-    @Override protected void chunkDone(double n){
+    @Override protected void chunkDone(long n){
       double r = 1.0/_nobs;
       _gram.mul(r);
     }
